moved predeformation script and also clean up the code following flake8
